[BLBX-6] refactoring : Separate session from video view model
